Refactored (re-arranged) route functions Renamed route files for clarity
