+++ http_server/flask_blueprints/html_sensor_configurations.py
@@ -1,13 +1,10 @@
 import os
 import shutil
 from flask import Blueprint, render_template, request
-
-import operations_modules.app_cached_variables
-import operations_modules.app_cached_variables_update
-import operations_modules.app_generic_functions
 from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules import app_cached_variables
+from operations_modules import app_cached_variables_update
 from operations_modules import app_generic_functions
 from operations_modules import app_config_access
 from operations_modules import network_ip
@@ -191,32 +188,6 @@ def html_edit_configurations():
                                message2="Unable to Display Configurations...")
 
 
-@html_sensor_config_routes.route("/HTMLRawConfigurations")
-@auth.login_required
-def view_https_config_diagnostics():
-    logger.network_logger.debug("** HTTPS Configuration Diagnostics accessed from " + str(request.remote_addr))
-    main_config = app_generic_functions.get_file_content(file_locations.main_config)
-    installed_sensors = app_generic_functions.get_file_content(file_locations.installed_sensors_config)
-    networking = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file)
-    wifi = app_generic_functions.get_file_content(file_locations.wifi_config_file)
-    trigger_variances = app_generic_functions.get_file_content(file_locations.trigger_variances_config)
-    sensor_control_config = app_generic_functions.get_file_content(file_locations.html_sensor_control_config)
-    weather_underground_config = app_generic_functions.get_file_content(file_locations.weather_underground_config)
-    luftdaten_config = app_generic_functions.get_file_content(file_locations.luftdaten_config)
-    open_sense_map_config = app_generic_functions.get_file_content(file_locations.osm_config)
-
-    return render_template("http_diagnostics_configurations.html",
-                           MainConfiguration=main_config,
-                           InstalledSensorsConfiguration=installed_sensors,
-                           NetworkConfiguration=networking,
-                           WiFiConfiguration=wifi,
-                           TriggerConfiguration=trigger_variances,
-                           SensorControlConfiguration=sensor_control_config,
-                           WeatherUndergroundConfiguration=weather_underground_config,
-                           LuftdatenConfiguration=luftdaten_config,
-                           OpenSenseMapConfiguration=open_sense_map_config)
-
-
 @html_sensor_config_routes.route("/EditConfigMain", methods=["POST"])
 @auth.login_required
 def html_set_config_main():
@@ -232,150 +203,6 @@ def html_set_config_main():
             logger.primary_logger.error("HTML Main Configuration set Error: " + str(error))
             return message_and_return("Bad Configuration POST Request",
                                                             url="/ConfigurationsHTML")
-
-
-@html_sensor_config_routes.route("/EditInstalledSensors", methods=["POST"])
-@auth.login_required
-def html_set_installed_sensors():
-    logger.network_logger.debug("** HTML Apply - Installed Sensors - Source " + str(request.remote_addr))
-    if request.method == "POST":
-        try:
-            check_html_installed_sensors(request)
-            app_generic_functions.thread_function(sensor_access.restart_services)
-            return message_and_return("Restarting Service, Please Wait ...",
-                                                            url="/ConfigurationsHTML")
-        except Exception as error:
-            logger.primary_logger.error("HTML Apply - Installed Sensors - Error: " + str(error))
-            return message_and_return("Bad Installed Sensors POST Request",
-                                                            url="/ConfigurationsHTML")
-
-
-@html_sensor_config_routes.route("/EditTriggerVariances", methods=["POST"])
-@auth.login_required
-def html_set_trigger_variances():
-    logger.network_logger.debug("** HTML Apply - Trigger Variances - Source " + str(request.remote_addr))
-    if request.method == "POST":
-        try:
-            check_html_variance_triggers(request)
-            return message_and_return("Trigger Variances Set",
-                                                            url="/ConfigurationsHTML")
-        except Exception as error:
-            logger.primary_logger.warning("HTML Apply - Trigger Variances - Error: " + str(error))
-    return message_and_return("Bad Trigger Variances POST Request",
-                                                    url="/ConfigurationsHTML")
-
-
-@html_sensor_config_routes.route("/ResetTriggerVariances")
-@auth.login_required
-def html_reset_trigger_variances():
-    logger.network_logger.info("** Trigger Variances Reset - Source " + str(request.remote_addr))
-    default_trigger_variances = CreateTriggerVariances()
-    app_config_access.trigger_variances = default_trigger_variances
-    write_triggers_to_file(default_trigger_variances)
-    return message_and_return("Trigger Variances Reset",
-                                                    url="/ConfigurationsHTML")
-
-
-@html_sensor_config_routes.route("/EditConfigIPv4", methods=["POST"])
-@auth.login_required
-def html_set_ipv4_config():
-    logger.network_logger.debug("** HTML Apply - IPv4 Configuration - Source " + str(request.remote_addr))
-    message2 = "Network settings have not been changed."
-    if request.method == "POST" and app_validation_checks.hostname_is_valid(request.form.get("ip_hostname")):
-        if request.form.get("ip_dhcp") is not None:
-            message2 = "You must reboot for all settings to take effect."
-            dhcpcd_template = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file_template)
-            dhcpcd_template = dhcpcd_template.replace("{{ StaticIPSettings }}", "")
-            hostname = request.form.get("ip_hostname")
-            os.system("hostnamectl set-hostname " + hostname)
-            app_generic_functions.write_file_to_disk(file_locations.dhcpcd_config_file, dhcpcd_template)
-            operations_modules.app_cached_variables_update.update_cached_variables()
-            return message_and_return("IPv4 Configuration Updated",
-                                                            text_message2=message2,
-                                                            url="/ConfigurationsHTML")
-
-        ip_address = request.form.get("ip_address")
-        ip_subnet = request.form.get("ip_subnet")
-        ip_gateway = request.form.get("ip_gateway")
-        ip_dns1 = request.form.get("ip_dns1")
-        ip_dns2 = request.form.get("ip_dns2")
-
-        try:
-            app_validation_checks.ip_address_is_valid(ip_address)
-        except ValueError:
-            return message_and_return("Invalid IP Address",
-                                                            text_message2=message2,
-                                                            url="/ConfigurationsHTML")
-        if not app_validation_checks.subnet_mask_is_valid(ip_subnet):
-            return message_and_return("Invalid Subnet Mask",
-                                                            text_message2=message2,
-                                                            url="/ConfigurationsHTML")
-        if ip_gateway is not "":
-            try:
-                app_validation_checks.ip_address_is_valid(ip_gateway)
-            except ValueError:
-                return message_and_return("Invalid Gateway",
-                                                                text_message2=message2,
-                                                                url="/ConfigurationsHTML")
-        if ip_dns1 is not "":
-            try:
-                app_validation_checks.ip_address_is_valid(ip_dns1)
-            except ValueError:
-                return message_and_return("Invalid Primary DNS Address",
-                                                                text_message2=message2,
-                                                                url="/ConfigurationsHTML")
-        if ip_dns2 is not "":
-            try:
-                app_validation_checks.ip_address_is_valid(ip_dns2)
-            except ValueError:
-                return message_and_return("Invalid Secondary DNS Address",
-                                                                text_message2=message2,
-                                                                url="/ConfigurationsHTML")
-
-        check_html_config_ipv4(request)
-        operations_modules.app_cached_variables_update.update_cached_variables()
-        return message_and_return("IPv4 Configuration Updated",
-                                                        text_message2="You must reboot the sensor to take effect.",
-                                                        url="/ConfigurationsHTML")
-    else:
-        return_message = "Invalid or Missing Hostname.\n\n" + \
-                         "Only Alphanumeric Characters, Dashes and Underscores may be used."
-        return message_and_return("Unable to Process IPv4 Configuration",
-                                                        text_message2=return_message,
-                                                        url="/ConfigurationsHTML")
-
-
-@html_sensor_config_routes.route("/EditConfigWifi", methods=["POST"])
-@auth.login_required
-def html_set_wifi_config():
-    logger.network_logger.debug("** HTML Apply - WiFi Configuration - Source " + str(request.remote_addr))
-    if request.method == "POST" and "ssid1" in request.form:
-
-        if app_validation_checks.text_has_no_double_quotes(request.form.get("wifi_key1")):
-            pass
-        else:
-            message = "Do not use double quotes in the Wireless Key Sections."
-            return message_and_return("Invalid Wireless Key",
-                                                            text_message2=message,
-                                                            url="/ConfigurationsHTML")
-
-        if app_validation_checks.wireless_ssid_is_valid(request.form.get("ssid1")):
-            new_wireless_config = check_html_config_wifi(request)
-            if new_wireless_config is not "":
-                return_message = "You must reboot the sensor to take effect."
-                operations_modules.app_cached_variables_update.update_cached_variables()
-                return message_and_return("WiFi Configuration Updated",
-                                                                text_message2=return_message,
-                                                                url="/ConfigurationsHTML")
-        else:
-            return_message = "Network Names cannot be blank and can only use " + \
-                             "Alphanumeric Characters, dashes, underscores and spaces."
-            return message_and_return("Unable to Process Wireless Configuration",
-                                                            text_message2=return_message,
-                                                            url="/ConfigurationsHTML")
-
-    return message_and_return("Unable to Process WiFi Configuration",
-                                                    url="/ConfigurationsHTML")
 
 
 def check_html_config_main(html_request):
@@ -411,6 +238,22 @@ def check_html_config_main(html_request):
         app_config_access.current_config.enable_custom_temp = 0
 
     app_config_access.config_primary.write_config_to_file(app_config_access.current_config)
+
+
+@html_sensor_config_routes.route("/EditInstalledSensors", methods=["POST"])
+@auth.login_required
+def html_set_installed_sensors():
+    logger.network_logger.debug("** HTML Apply - Installed Sensors - Source " + str(request.remote_addr))
+    if request.method == "POST":
+        try:
+            check_html_installed_sensors(request)
+            app_generic_functions.thread_function(sensor_access.restart_services)
+            return message_and_return("Restarting Service, Please Wait ...",
+                                                            url="/ConfigurationsHTML")
+        except Exception as error:
+            logger.primary_logger.error("HTML Apply - Installed Sensors - Error: " + str(error))
+            return message_and_return("Bad Installed Sensors POST Request",
+                                                            url="/ConfigurationsHTML")
 
 
 def check_html_installed_sensors(html_request):
@@ -476,6 +319,21 @@ def check_html_installed_sensors(html_request):
 
     installed_sensors_text = new_installed_sensors.get_installed_sensors_config_as_str()
     app_config_access.config_installed_sensors.write_to_file(installed_sensors_text)
+
+
+@html_sensor_config_routes.route("/EditTriggerVariances", methods=["POST"])
+@auth.login_required
+def html_set_trigger_variances():
+    logger.network_logger.debug("** HTML Apply - Trigger Variances - Source " + str(request.remote_addr))
+    if request.method == "POST":
+        try:
+            check_html_variance_triggers(request)
+            return message_and_return("Trigger Variances Set",
+                                                            url="/ConfigurationsHTML")
+        except Exception as error:
+            logger.primary_logger.warning("HTML Apply - Trigger Variances - Error: " + str(error))
+    return message_and_return("Bad Trigger Variances POST Request",
+                                                    url="/ConfigurationsHTML")
 
 
 def check_html_variance_triggers(html_request):
@@ -650,6 +508,140 @@ def check_html_variance_triggers(html_request):
     write_triggers_to_file(app_config_access.trigger_variances)
 
 
+@html_sensor_config_routes.route("/ResetTriggerVariances")
+@auth.login_required
+def html_reset_trigger_variances():
+    logger.network_logger.info("** Trigger Variances Reset - Source " + str(request.remote_addr))
+    default_trigger_variances = CreateTriggerVariances()
+    app_config_access.trigger_variances = default_trigger_variances
+    write_triggers_to_file(default_trigger_variances)
+    return message_and_return("Trigger Variances Reset", url="/ConfigurationsHTML")
+
+
+@html_sensor_config_routes.route("/EditConfigIPv4", methods=["POST"])
+@auth.login_required
+def html_set_ipv4_config():
+    logger.network_logger.debug("** HTML Apply - IPv4 Configuration - Source " + str(request.remote_addr))
+    message2 = "Network settings have not been changed."
+    if request.method == "POST" and app_validation_checks.hostname_is_valid(request.form.get("ip_hostname")):
+        if request.form.get("ip_dhcp") is not None:
+            message2 = "You must reboot for all settings to take effect."
+            dhcpcd_template = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file_template)
+            dhcpcd_template = dhcpcd_template.replace("{{ StaticIPSettings }}", "")
+            hostname = request.form.get("ip_hostname")
+            os.system("hostnamectl set-hostname " + hostname)
+            app_generic_functions.write_file_to_disk(file_locations.dhcpcd_config_file, dhcpcd_template)
+            app_cached_variables_update.update_cached_variables()
+            return message_and_return("IPv4 Configuration Updated", text_message2=message2, url="/ConfigurationsHTML")
+
+        ip_address = request.form.get("ip_address")
+        ip_subnet = request.form.get("ip_subnet")
+        ip_gateway = request.form.get("ip_gateway")
+        ip_dns1 = request.form.get("ip_dns1")
+        ip_dns2 = request.form.get("ip_dns2")
+
+        try:
+            app_validation_checks.ip_address_is_valid(ip_address)
+        except ValueError:
+            return message_and_return("Invalid IP Address",
+                                                            text_message2=message2,
+                                                            url="/ConfigurationsHTML")
+        if not app_validation_checks.subnet_mask_is_valid(ip_subnet):
+            return message_and_return("Invalid Subnet Mask",
+                                                            text_message2=message2,
+                                                            url="/ConfigurationsHTML")
+        if ip_gateway is not "":
+            try:
+                app_validation_checks.ip_address_is_valid(ip_gateway)
+            except ValueError:
+                return message_and_return("Invalid Gateway",
+                                                                text_message2=message2,
+                                                                url="/ConfigurationsHTML")
+        if ip_dns1 is not "":
+            try:
+                app_validation_checks.ip_address_is_valid(ip_dns1)
+            except ValueError:
+                return message_and_return("Invalid Primary DNS Address",
+                                                                text_message2=message2,
+                                                                url="/ConfigurationsHTML")
+        if ip_dns2 is not "":
+            try:
+                app_validation_checks.ip_address_is_valid(ip_dns2)
+            except ValueError:
+                return message_and_return("Invalid Secondary DNS Address",
+                                                                text_message2=message2,
+                                                                url="/ConfigurationsHTML")
+
+        check_html_config_ipv4(request)
+        app_cached_variables_update.update_cached_variables()
+        return message_and_return("IPv4 Configuration Updated",
+                                                        text_message2="You must reboot the sensor to take effect.",
+                                                        url="/ConfigurationsHTML")
+    else:
+        return_message = "Invalid or Missing Hostname.\n\n" + \
+                         "Only Alphanumeric Characters, Dashes and Underscores may be used."
+        return message_and_return("Unable to Process IPv4 Configuration",
+                                                        text_message2=return_message,
+                                                        url="/ConfigurationsHTML")
+
+
+def check_html_config_ipv4(html_request):
+    logger.network_logger.debug("Starting HTML IPv4 Configuration Update Check")
+    dhcpcd_template = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file_template)
+
+    hostname = html_request.form.get("ip_hostname")
+    os.system("hostnamectl set-hostname " + hostname)
+    app_cached_variables.hostname = hostname
+
+    ip_address = html_request.form.get("ip_address")
+    ip_subnet_mask = html_request.form.get("ip_subnet")
+    ip_gateway = html_request.form.get("ip_gateway")
+    ip_dns1 = html_request.form.get("ip_dns1")
+    ip_dns2 = html_request.form.get("ip_dns2")
+
+    ip_network_text = "interface wlan0\nstatic ip_address=" + ip_address + ip_subnet_mask + "\nstatic routers=" + ip_gateway + \
+                      "\nstatic domain_name_servers=" + ip_dns1 + " " + ip_dns2
+
+    dhcpcd_template = dhcpcd_template.replace("{{ StaticIPSettings }}", ip_network_text)
+    network_ip.write_ipv4_config_to_file(dhcpcd_template)
+
+    shutil.chown(file_locations.dhcpcd_config_file, "root", "netdev")
+    os.chmod(file_locations.dhcpcd_config_file, 0o664)
+
+
+@html_sensor_config_routes.route("/EditConfigWifi", methods=["POST"])
+@auth.login_required
+def html_set_wifi_config():
+    logger.network_logger.debug("** HTML Apply - WiFi Configuration - Source " + str(request.remote_addr))
+    if request.method == "POST" and "ssid1" in request.form:
+
+        if app_validation_checks.text_has_no_double_quotes(request.form.get("wifi_key1")):
+            pass
+        else:
+            message = "Do not use double quotes in the Wireless Key Sections."
+            return message_and_return("Invalid Wireless Key",
+                                                            text_message2=message,
+                                                            url="/ConfigurationsHTML")
+
+        if app_validation_checks.wireless_ssid_is_valid(request.form.get("ssid1")):
+            new_wireless_config = check_html_config_wifi(request)
+            if new_wireless_config is not "":
+                return_message = "You must reboot the sensor to take effect."
+                app_cached_variables_update.update_cached_variables()
+                return message_and_return("WiFi Configuration Updated",
+                                                                text_message2=return_message,
+                                                                url="/ConfigurationsHTML")
+        else:
+            return_message = "Network Names cannot be blank and can only use " + \
+                             "Alphanumeric Characters, dashes, underscores and spaces."
+            return message_and_return("Unable to Process Wireless Configuration",
+                                                            text_message2=return_message,
+                                                            url="/ConfigurationsHTML")
+
+    return message_and_return("Unable to Process WiFi Configuration",
+                                                    url="/ConfigurationsHTML")
+
+
 def check_html_config_wifi(html_request):
     logger.network_logger.debug("Starting HTML WiFi Configuration Update Check")
     if html_request.form.get("ssid1") is not None:
@@ -684,25 +676,27 @@ def check_html_config_wifi(html_request):
         return ""
 
 
-def check_html_config_ipv4(html_request):
-    logger.network_logger.debug("Starting HTML IPv4 Configuration Update Check")
-    dhcpcd_template = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file_template)
+@html_sensor_config_routes.route("/HTMLRawConfigurations")
+@auth.login_required
+def html_raw_configurations_view():
+    logger.network_logger.debug("** HTML Raw Configurations viewed by " + str(request.remote_addr))
+    main_config = app_generic_functions.get_file_content(file_locations.main_config)
+    installed_sensors = app_generic_functions.get_file_content(file_locations.installed_sensors_config)
+    networking = app_generic_functions.get_file_content(file_locations.dhcpcd_config_file)
+    wifi = app_generic_functions.get_file_content(file_locations.wifi_config_file)
+    trigger_variances = app_generic_functions.get_file_content(file_locations.trigger_variances_config)
+    sensor_control_config = app_generic_functions.get_file_content(file_locations.html_sensor_control_config)
+    weather_underground_config = app_generic_functions.get_file_content(file_locations.weather_underground_config)
+    luftdaten_config = app_generic_functions.get_file_content(file_locations.luftdaten_config)
+    open_sense_map_config = app_generic_functions.get_file_content(file_locations.osm_config)
 
-    hostname = html_request.form.get("ip_hostname")
-    os.system("hostnamectl set-hostname " + hostname)
-    app_cached_variables.hostname = hostname
-
-    ip_address = html_request.form.get("ip_address")
-    ip_subnet_mask = html_request.form.get("ip_subnet")
-    ip_gateway = html_request.form.get("ip_gateway")
-    ip_dns1 = html_request.form.get("ip_dns1")
-    ip_dns2 = html_request.form.get("ip_dns2")
-
-    ip_network_text = "interface wlan0\nstatic ip_address=" + ip_address + ip_subnet_mask + "\nstatic routers=" + ip_gateway + \
-                      "\nstatic domain_name_servers=" + ip_dns1 + " " + ip_dns2
-
-    dhcpcd_template = dhcpcd_template.replace("{{ StaticIPSettings }}", ip_network_text)
-    network_ip.write_ipv4_config_to_file(dhcpcd_template)
-
-    shutil.chown(file_locations.dhcpcd_config_file, "root", "netdev")
-    os.chmod(file_locations.dhcpcd_config_file, 0o664)
+    return render_template("view_raw_configurations.html",
+                           MainConfiguration=main_config,
+                           InstalledSensorsConfiguration=installed_sensors,
+                           NetworkConfiguration=networking,
+                           WiFiConfiguration=wifi,
+                           TriggerConfiguration=trigger_variances,
+                           SensorControlConfiguration=sensor_control_config,
+                           WeatherUndergroundConfiguration=weather_underground_config,
+                           LuftdatenConfiguration=luftdaten_config,
+                           OpenSenseMapConfiguration=open_sense_map_config)
