+++ http_server/flask_blueprints/text_sensor_readings.py
@@ -6,6 +6,13 @@ from sensor_modules import sensor_access
 html_sensor_readings_routes = Blueprint("html_sensor_readings_routes", __name__)
 
 
+# This one isn't in use yet, but there for allowing one sensor to record all readings
+@html_sensor_readings_routes.route("/GetIntervalSensorReadings")
+def get_interval_readings():
+    logger.network_logger.debug("* Interval Sensor Readings sent to " + str(request.remote_addr))
+    return str(sensor_access.get_interval_sensor_readings())
+
+
 @html_sensor_readings_routes.route("/GetCPUTemperature")
 def get_cpu_temperature():
     logger.network_logger.debug("* Sensor's CPU Temperature sent to " + str(request.remote_addr))
