+++ http_server/templates/view_raw_configurations.html
